+++ src/app/app-routing.module.ts
@@ -14,6 +14,27 @@ const routes: Routes = [
     loadChildren: () => import('./pages/about/about.module').then( m => m.AboutPageModule)
   },
   {
+    path: 'cookie-policy',
+    data: {
+      backendPageId: '05-01'
+    },
+    loadChildren: () => import('./pages/about/about.module').then( m => m.AboutPageModule)
+  },
+  {
+    path: 'privacy-policy',
+    data: {
+      backendPageId: '05-02'
+    },
+    loadChildren: () => import('./pages/about/about.module').then( m => m.AboutPageModule)
+  },
+  {
+    path: 'terms',
+    data: {
+      backendPageId: '05-03'
+    },
+    loadChildren: () => import('./pages/about/about.module').then( m => m.AboutPageModule)
+  },
+  {
     path: 'article',
     loadChildren: () => import('./pages/article/article.module').then( m => m.ArticlePageModule)
   },